ml_project/requirements.txt
attrs==21.2.0
boto3==1.17.69
botocore==1.20.69
click==7.1.2
iniconfig==1.1.1
jmespath==0.10.0
joblib==1.0.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
typeguard==2.12.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
